post and put for bands tuned and working

+++ src/main/java/com/java/bandify/persistance/db/repository/BandRepository.java
@@ -3,6 +3,6 @@ package com.java.bandify.persistance.db.repository;
 import com.java.bandify.persistance.db.entity.BandEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BandRepository extends JpaRepository<BandEntity, Long> {
+public interface BandRepository extends JpaRepository<BandEntity, Integer> {
 
 }
