+++ src/main/java/com/java/bandify/persistance/db/repository/CountryRepository.java
@@ -4,6 +4,6 @@ import com.java.bandify.persistance.db.entity.BandEntity;
 import com.java.bandify.persistance.db.entity.CountryEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CountryRepository extends JpaRepository<CountryEntity, Long> {
+public interface CountryRepository extends JpaRepository<CountryEntity, Integer> {
 
 }
