+++ src/main/java/com/java/bandify/persistance/db/repository/CountyRepository.java
@@ -4,6 +4,6 @@ import com.java.bandify.persistance.db.entity.BandEntity;
 import com.java.bandify.persistance.db.entity.CountyEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CountyRepository extends JpaRepository<CountyEntity, Long> {
+public interface CountyRepository extends JpaRepository<CountyEntity, Integer> {
 
 }
