+++ src/main/java/com/java/bandify/persistance/db/repository/CurrencyRepository.java
@@ -1,9 +1,8 @@
 package com.java.bandify.persistance.db.repository;
 
-import com.java.bandify.persistance.db.entity.BandEntity;
 import com.java.bandify.persistance.db.entity.CurrencyEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CurrencyRepository extends JpaRepository<CurrencyEntity, Long> {
+public interface CurrencyRepository extends JpaRepository<CurrencyEntity, Integer> {
 
 }
