+++ src/main/java/com/java/bandify/persistance/db/repository/GenreRepository.java
@@ -4,6 +4,6 @@ import com.java.bandify.persistance.db.entity.BandEntity;
 import com.java.bandify.persistance.db.entity.GenreEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GenreRepository extends JpaRepository<GenreEntity, Long> {
+public interface GenreRepository extends JpaRepository<GenreEntity, Integer> {
 
 }
