+++ src/main/java/com/java/bandify/persistance/db/repository/InstrumentRepository.java
@@ -3,6 +3,6 @@ package com.java.bandify.persistance.db.repository;
 import com.java.bandify.persistance.db.entity.InstrumentEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InstrumentRepository extends JpaRepository<InstrumentEntity, Long> {
+public interface InstrumentRepository extends JpaRepository<InstrumentEntity, Integer> {
 
 }
