+++ src/main/java/com/java/bandify/persistance/db/repository/MusicianRepository.java
@@ -4,6 +4,6 @@ import com.java.bandify.persistance.db.entity.BandEntity;
 import com.java.bandify.persistance.db.entity.MusicianEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MusicianRepository extends JpaRepository<MusicianEntity, Long> {
+public interface MusicianRepository extends JpaRepository<MusicianEntity, Integer> {
 
 }
