+++ src/main/java/com/java/bandify/persistance/db/repository/NonmusicianRepository.java
@@ -4,6 +4,6 @@ import com.java.bandify.persistance.db.entity.BandEntity;
 import com.java.bandify.persistance.db.entity.NonmusicianEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NonmusicianRepository extends JpaRepository<NonmusicianEntity, Long> {
+public interface NonmusicianRepository extends JpaRepository<NonmusicianEntity, Integer> {
 
 }
