+++ src/main/java/com/java/bandify/persistance/db/repository/TownRepository.java
@@ -4,6 +4,6 @@ import com.java.bandify.persistance.db.entity.BandEntity;
 import com.java.bandify.persistance.db.entity.TownEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TownRepository extends JpaRepository<TownEntity, Long> {
+public interface TownRepository extends JpaRepository<TownEntity, Integer> {
 
 }
